Simplifying logic at the cost of less logs
We can check if the sortition pool is locked at the very beginning and
return if it is. It will simplify the logic, however we will not log if
the operator is in the pool or if it's up to date. This commit can be
reverted if we decide that we need additional loggin.