Improved the implementation of reward eligibility local chain functions
Instead of keeping a zero-value entry in the mapping for ineligibleUntil we just
delete the entire mapping. This approach is more Go-style and allows to simplify
the code.

The original implementation reverts for an operator that is already eligible and
tries restore eligibility. We now mimic this behaviour in the local
implementation.

Last but not least we now acquire the lock once for the entire execution of
RestoreRewardEligibility. The previous version was not thread safe: someone
could mark operator as ineligible for rewards after the read lock release.