Improvements to local sortition chain unit tests
Three small improvements were made in this commit.

First, given the operator is eligible for rewards by default, we can drop the
line setting reward ineligibility to zero.

Second, we use non-zero value for `ineligibleUntil` when testing. Ineligible
until 0 does not really make sense.

Lastly, the assertion for RestoreRewardEligibility is stronger now given we
assert the particular error is returned.