Deleted eligible for rewards pass-thru unit test
There are no changes applied by `MonitorPool` and we are not asserting any
changes were applied. If we remove the `MonitorPool(ctx, localChain, statusCheckTick)`
call, this test is still passing.

An operator is eligible for rewards by default. If there was something wrong
with the eligibility check, most of the previous tests would fail.