Outline of the `pkg/ecdsa` and `pkg/tbtc`
This changeset introduces an outline of two
new packages.

The `pkg/ecdsa` will hold the code of an ECDSA
node. This package will be responsible for the
interaction with the on-chain ECDSA contracts,
will manage the lifecycle of the ECDSA
wallets, and will provide a way to obtain ECDSA
signatures. The package is designed with
re-usability in mind so, it should be possible
to use it by various client applications if
needed.

The `pkg/tbtc` will hold the code of the TBTC
application that leverages the `pkg/ecdsa`
package for signing Bitcoin transactions. This
package will be responsible for the integration
with the on-chain `Bridge` contract and with the
Bitcoin network. It will also manage ECDSA
wallets in the scope required by the Bridge.
The package is greatly inspired by the
existing `pkg/beacon` code though a lot of
simplifications and improvements will be made.