Restore old keep-core v1 deploy script
There was a change recently that refactored the `install.sh` script in
such a way that it's deploying the contracts on local geth instead of
ganache. This caused some problem when deploying the contracts locally
for ganache when developing the threshold dApp.

We are temporarly restoring back the old keep-core v1 deploy script as a
new file to get rid of errors in deployment proccess and pretty soon we
will refactor the whole process and move to goerli testnet (instead of
developing the contracts locally).