Support multiple accounts for hardhat test network
We want to be able to configure private keys to multiple accounts for
example on goerli network to be able to run initialization scripts.
With this change we expect the environment property to define one or
more private keys separated with a comma (`,`).

The change is made also to the variable name, to better reflect the
meaning.