Define developer flags for contract addresses
We want to be able to overwrite default contract addresses with flags
and config files, option for this should be
`--developer.<contractName>Address` e.g.
`--developer.randomBeaconAddress`.

To do that we need to register flags for a command and bind them with
config.

The tricky part is that `config.Config` struct keeps the addresses in
`Ethereum.ContractAddresses` map, which is a different key. We need to
alias the `developer.<contractName>Address` key with
`Ethereum.ContractAddresses.<ContractName>` property in Config struct to
get it working.