Changing indexes of the named accounts & flag for external contracts
We need to align with the ecdsa named accounts. Whenever installation
script deploys beacon and ecdsa contracts, some of the contracts are
being approved in another contracts by the owner/deployer. The deployer
has to be the same for the local development purposes.

Added a flag to use the external deploy. The external contracts is
usefull for the all-in-one approach for hardhat tests.