Resolve default addresses for contracts in config
We don't want to resolve the default addresses in developer flags as
this may be confusing and the default address is printed in the help
message.

The addresses are resolved in the config, after reading flags and config
file. If addresses are not set we're setting them to defaults. This is
preparation for the multi-network client, where we will resolve
contracts addresses based on a `--network` flag provided for the client
run.