Move download_artifacts target to the main Makefile
We moved the artifacts downloading to a main Makefile.
Now it's enough to run `make` in the root directory to download the
artifacts and build the client.

To run make for sepcific network(environment) the command should be
executed as, e.g.
make environment=goreli

No need to set the CONTRACTS_NPM_PACKAGE_TAG env variable.