Add targets for each environment
It's now simpler to run make for specific environment, the commands are:
`make development`, `make goerli`, in the future we will have `make
mainnet`.

For now the default command (the first one in the Makefile is for
development, so it's enough to run just `make` to build client with
development environmnet.