Introduce basic domain types
Here we introduce the basic domain types for
the tBTC application. The most important of them
is the `tbtc.wallet` type which represents a
tBTC wallet. A `tbtc.wallet` is an identifiable
set of `signer`s which are able to perform
the tECDSA signing using their underlying
`tecdsa.PrivateKeyShare`s.