Added extra gas requirement for ECDSA DKG result challenge
Due to EIP150, 1/64 of the gas is not forwarded to the call, and will be kept to
execute the remaining operations in the function after the call inside the
try-catch.

To ensure there is no way for the caller to manipulate the gas limit in such
a way that the call inside try-catch fails with out-of-gas and the rest of the
function is executed with the remaining 1/64 of gas, we require an extra gas
amount to be left at the end of the call to `WalletRegistry.challengeDkgResult`.

In practice, with the current version of the code, such a scenario should not be
possible anyway but the extra gas requirement provides an extra layer of
protection in case of any code changes. In the current version of the code, the
cost of `dkg.challegeResult(dkgResult)` is about 1.6M gas units and the cost of
everything that happens later is around 144k gas units. It means that the cost
of the rest of the logic is about 6% of the cost of the entire function which is
more than 1/64.

Another option for solving this problem was trying to eliminate try-catch
altogether. This option would be more error-prone for future changes in the code
though. We would need to check all functions on the execution stack of the DKG
result challenge and ensure they can not revert or if they revert, they revert
with a specific error. Any future change to the code would require inspecting
the possible cases of reverts again. That's why a general try-catch with an
extra gas requirement at the end feels safer long-term.

Note that the functions on the execution stack of the DKG result challenge may
revert. One example is the corrupted signature verification covered by
`with dkg result submitted with unrecoverable signatures` unit tests. That's why
some try-catch is necessary, even if not on the highest level, then more
try-catches on the lower levels of the execution stack.