Adjust `client.yml` for use in v2
In v2 of the system we modified the structure of the components that we
build to set up the system. There have been changes in the structure of
the `keep-core` repository and the order of the execution of modules in
the CI has changed as well. In this PR/commit we upgrade the custom CI
actions to `v2`, which uses the latest module order. We also adjust
slightly the workflow triggering conditions.