Added pre-params generation concurrency to ExecutorConfig
We need to be able to configure TSS pre-params generation concurrency to
not spend the entire computing power on just TSS pre-parameters,
starving the rest of the client logic.

This commit is just a change to `ExecutorConfig` and makes the
concurrency set to 1. In the next commits, this property and others will
be made configurable.