Allow configuring tECDSA  parameters for the client
It is now possible to configure three parameters of tECDSA protocol:
- pre-params pool size
- pre-params generation timeout
- pre-params generation concurrency

The parameters can be configured via config file or via command-line flags.

Worth noting, that in tss-lib, concurrency level < 3 translates to a concurrency
level equal to 1:

From v1.3.3 of `tss-lib/ecdsa/keygen/prepare.go`:
```
if concurrency /= 3; concurrency < 1 {
concurrency = 1
}
```
https://github.com/bnb-chain/tss-lib/blob/v1.3.3/ecdsa/keygen/prepare.go#L41-L43

This is because there are three concurrent generation threads triggered in
`prepare.go`.