Move pre-parameters config to tbtc
If we have things like `PreParamsPoolSize` right in `pkg/tecdsa` that means this
package is aware that there is something like a pre params pool in the DKG
implementation. That information leaks from `pkg/tecdsa/dkg` and it is
`pkg/tecdsa/dkg` that uses `pkg/tecdsa`, not vice-versa. If this was about
something really general for tECDSA like, for example, threshold or so,
everything would be fine. The `pkg/tbtc` is the application and it should be
responsible for its own config (i.e. should declare it). Since it uses
`pkg/tecdsa/dkg` directly, the knowledge about pre-params pool in `pkg/tbtc`
seems to be justified, especially since it initiates the executor passing
parameters from the config.