Abstraction to manage computationally-heavy operation
Miner allows managing computationally heavy operations: stopping and resuming
them. The client needs to generate parameters for cryptographic algorithms
and generating these parameters requires a lot of CPU cycles. The generation
process may starve other processes in the client when it comes to access to
the CPU. The miner allows starting the parameter generation when no other
processes such as key generation or signing are executed on the client. This
way, the client that would normally be idle, can spend CPU cycles on
computationally heavy operations and stop these operations when CPU cycles
are needed elsewhere.