Use separate preparams pb type for LocalPreParams and LocalPartySaveData
In 9c62fdd I promoted LocalPreParams to be defined outside of
LocalPartySaveData. With this change, in 78dc79f, I could move LocalPreParams
message to a separate file. This message is used by LocalPartySaveData and by
PreParams that will be saved on disk.

I was struggling with PrivateKey part: it is empty for PreParams (before DKG)
and is filled for LocalPartySaveData.

After consulting with @lukasz-zimnoch we decided to  keep LocalPreParams as
internal for LocalPartySaveData and introduce separate LocalPreParams without
PrivateKey and use it as an internal message of PreParams.

A little duplication seems to be a much smaller problem than introducing the
additional complexity related to a field that can be in different states
depending on the usage context. Protobuf definitions are meant to describe the
shape of data and we think we should not have any details regarding how those
data are processed (i.e. this field should be empty before DKG but filled after).
Additionally, the structure with re-used pre-params can be error-prone,
especially regarding the marshaling code.