Introduced ParameterPool abstraction
To have the generated TSS pre-parameters survive client restarts, we need to add
a persistence layer to the pool. The abstraction allows to define and test all
functionalities separately, without having to worry about the complexity of the
generation of real parameters.

This way, TSS pre-parameter code can be just wiring things up: take the
generation function, take the persistence layer, provide them to the introduced
abstraction, and voila!

The complexity of the pool logic can be separated from the complexity of wiring
things up.