Use ParameterPool abstraction for TSS pre-parameters
In the previous commit, an abstraction for generating parameters have been
introduced. In this commit, the TSS pre-parameters code is updated to use the
abstraction.

This way, TSS pre-parameter code can be just wiring things up: take the
generation function, take the persistence layer, provide them to the introduced
abstraction, and voila! The complexity of the pool logic can be separated from
the complexity of wiring things up.