Handle long-running computations in tECDSA DKG states
Here we are refactoring tECDSA DKG states that
perform long-running TSS computations. As those
computations may take longer than the state's
lifetime, we need to set up a proper timeout
signal to avoid resource leaks. The state's
context passed as argument of the `state.Initiate`
function would be a good candidate, but it works
as expected only if `state.Initiate` returns
before the state's end block is reached.

So far, we were using a simple fixed timeout based
on state blocks duration though this solution is
not sufficient for the long term as the average
block time can change and lead to weird corner
cases.

Instead of that, we are now triggering
long-running computations and message broadcasting
in a separate goroutine upon state initiation
and use the state's context as the timeout
signal. In order to handle the state logic's
outcome, we are using the success and errors
channels that pipe the outcome to the
`state.Next` function which is invoked at the
end of the state. If state's computations were
successful, we're proceeding to the next state.
Otherwise, a state error bubbles up and breaks
the entire state machine instance.

The aforementioned mechanism requires some changes
in the state machine logic and `state.State`
interface that are also part of this changeset.