Cancel the context of worker function when miner is stopped
This change allows stopping the worker function immediately when the miner
receives the stop signal. This is important for long-running worker functions
when a single iteration can take a couple of minutes. Thanks to the change from
tss-lib (https://github.com/bnb-chain/tss-lib/pull/191) we can now stop the
pre-param generation immediately with this mechanism.

I have also added a mutex to `mockPersistence` in `pool_test.go`. There is
a single goroutine writing to the map, but it's a separate goroutine from the
test one reading from a map so having a mutex feels safer, especially since
there might be some changes in miner or some other tests can be added to
`pool_test.go`.