Move Categories definition to config package to use in validation
We move the categories to the config package to use them for config
validation.
It may happen (e.g. for ethereum command) that we want to use just a
specific category of configuration, we cannot expect that in such case
validation for all parameters is performed as only those related to
specific category will be provided.