Improved TestPersist unit test increasing the pool size
The test failed on CI but keeps passing locally. There was a potential
problem when the pool generated enough parameters to reach the capacity
and although all of them were persisted, the last one was not written to
the buffered channel given it's blocking operation. In this case
pool.CurrentSize() is off-by-one to persistence.parameterCount().