Hardhat task for requesting new ECDSA wallets
Here we introduce a Hardhat task that allows requesting for a new ECDSA wallet.
This task should facilitate testing on local Ethereum networks.
In order to run it, move to the `solidity/ecdsa` directory and do:
```
npx hardhat request-new-wallet --wallet-owner-address <owner> --network
<network-name>
```
where `owner` is the wallet owner's address that can be set using the
`initialize-wallet-owner` Hardhat task.