Redefined the parameter generator API
The first change is the series of renames and function export changes. To do not
confuse our operators that keep client is "mining" blocks, the miner package and
struct have been renamed to `generator` and `Scheduler`. The `Scheduler` name
better describes the responsibilities of the structure - it is stopping and
resuming long-running tasks.

Secondly, we no longer expect protocol goroutines to call `Stop()` and
`Resume()`. Instead, there is a global monitoring goroutine started when the
Scheduler is created that queries the protocols. This change allows to remove
the quite complicated concept of a WaitGroup-based latch and will allow to not
pass the Scheduler instance down to all goroutines.

The implementations of Protocols will be done in the next commits. For now, no
protocols are registered when Scheduler is started.