Simple latch to control the state of the protocol
ProtocolLatch increases the internal counter every time protocol execution
starts and decreases the counter every time protocol execution completes.
The latch implements Protocol interface and can be registered in the
Scheduler.

The protocol code using the latch must guarantee that:
1. `Lock()` is always called before `Unlock()`
2. `Unlock()` is eventually called for every `Lock()`.

Note that the Unlock() function may panic if the conditions are not met.

The latch is not yet used - it is going to be integrated with protocols in the
next commits.