Use forked sorition pools in testnet deployment
We want to modify sortiion pools used in test environment to make
joining applications easier for operators. With the changes from forked
sortition pools we will not require sortition pool to be unlocked when
operator wants to join. Even in case the DKG process is in progress we
will let operators to join.

This is just a temporary change that should be reverted in the near
future.

Refs https://github.com/keep-network/sortition-pools/pull/186