Initialize ethereum network boolean flags
The flags were added to specify which network the client should connect
to. The idea behind the flags is that the same build of the client will
be able to connect to different networks.
To run the client with Goerli add `--goerli` flag, to run the client
with local developer network add `--developer` flag. If non of the
afromentioned flags is used the client will default to the Mainnet.

The `--mainnet` flag is defined only temporary to reduce the confusion
when developing the client.