Properly map DKG result to a wallet signing group
The final signing group established at the end
of the DKG protocol may differ from the original
group outputted by the sortition protocol.
The final group should consist from members
who behaved correctly during the DKG protocol
and final members indexes should reflect
the position of particular members in that group
in a correct way. This is not only important
for code consistency but is also required by
the `WalletRegistry` contract where functions
like `isWalletMember` rely on properly built
wallet members hash.

This changeset applies several changes that
adjust the code to the aforementioned requirements
and improves the documentation, logging, and test
coverage around it.