Simplify DKG result signature verification function
This phase also requires that there is only
one message per given sender. To ensure that,
we can use the `deduplicateBySender` function
we added recently.

Apart from that, there is no need to check for
message from self in this function. This
is ensured one level above, when the message
is received and validated using
`shouldAcceptMessage` function.