Validate staking contract addresses
It is expected that TokenStaking contract from threshold module is
used by the beacon (RandomBeacon) and ecdsa/tbtc (WalletRegistry).
All three modules should use the same instance of the contract.

During development testing it caused a lot of debugging to figure out the
problem with the client that was caused by beacon and ecdsa contracts
packages using an old threshold contracts.

Here we add a simple validation to check if the TokenStaking used by
beacon and tbtc match the base chain.