Validate Chain ID on ethereum connection establishement
We expect the client to be run for a network specified with boolean
flags like `--mainnet` and `--goerli`. Another setting that is required
is URL to the ethereum API URL.
Here we want to ensure that the API URL works with the network that the
client is expected to be run against. We use Chain ID for validation.

The exception is `--developer` network, that we won't validate the Chain
ID as it can vary.