Add pagination to subgraph queries
Since there is a 1000 elements limit per entity to data retrieved by a
subgraph GraphQL query, if in any moment any query reach this limit,
elements exceeding this limit will be missed, and the calculation of
rewards for merkle distribution will not be correct.

The proposed solution here is to retrieves all the elements in batchs
of 1000 elements.