Allow configuring key generation concurrency
By default, key generation concurrency is set to the number of cores available
on the machine. This may not be an optimal choice for multiple clients running
on the same machine, especially if they have a lot of group members. It is also
possible that the same machine is running other services (Ethereum client,
Electrum node), and occupying all cores of the CPU with a long-running DLN proof
verification is not desirable. This change allows configuring key generation
concurrency level for TBTC. If set to N, no more than N cores are used for
long-running TSS DLN proof verification during the key generation.