Address two TODOs about `GroupQuorum` param
The first one about `len(operatingMembersIndexes)
< chainConfig.GroupQuorum` is obvious but
regarding the second one, it looks like we need
to compare the number of DKG result sigs
to the honest threshold. This number is enough,
and we don't need the group quorum to sign
the result as this is not expected by the on-chain
contract as well.