Add workflow building and publishing docs from `docs` folder
The workflow replaces the old flow that was used to build `v1`-related documents
(`docs-v1.yml`). The `v1` docs are no longer relevant, we now need to publish
the documents related to `v2`. The workflow is set to trigger automatically when
there are changes to files in `docs` directory.
When changes are part of PR, the docs will be deployed to yhe preview bucket.
When changes are part of merge to master (PR merge), the docs will be deployed
to the main bucket.