Provide a cancel signal for the parameter pool getter
The `Get` function exposed by the `generator
.ParameterPool` type can hang on the
`<-pp.pool` instruction in case the parameter
pool is empty. Although this behavior is proper
if the pool is working, it may cause deadlocks
if the pool is stopped by the scheduler who
is waiting for resume signals from clients
which are in turn waiting for
parameters that are supposed to be provided by the
empty pool. Such a scenario occurred for tECDSA
DKG. Here we aim to fix it by providing a context
argument to the `Get` function.