Add comment to PR-s modifying docs that links to the published preview
We're adding a step to the `docs.yml` workflow that publishes the comment under
the PR modifying the documents, after preview of the dosc gets published to the
GCP bucket. The comment contains a link to the `index.html` page containing the
table of contents.
Note: we need to have `./docs/index.adoc` file with right config inside to have
it working in the way I described it.