Make common ethereum flags persistent
The flags are required by `validateConfig` to be set. Also we use them
in the ethereum comamnds in the generated contract commands to
initialize ethereum connection for contracts go code, we need to define
them persisten on `ethereum` command level, so they are avialable in the
chikd subcommands for contracts functions execution.