Rework `generator.ParameterPool` getter
We decided to not use the timeout and use a more
suitable logic that returns a parameter or
an error immediately. This is because there
is no sense to wait for new parameters when the
pool is empty as the scheduler is typically
stopped when parameters are taken from the pool.