Yield the processor on each test loop execution
We yield the processor with `runtime.Gosched()` to ensure the processor
is not stuck for too long in the pool's current size check loop and has
enough computing power to generate pre-parameters.

For most of the time, it should not be a problem but we could have this
test timing out from time to time in extreme circumstances.