Add Private Testnet bundles generator
We need to generate a bundles for the parties that are joining our
Private Testnet.

The generated bundles contains a preinitialized Ethereum Account details. The account
receives a stake delegation to the `provider` account with authorized `beacon` and
`tbtc` application. This is a production-like experience for the Staking Providers,
where they will receive stakes from their customers.

We have a scripts to create a bundle and initialize the provider
account.

Please see README.md for more details.