Adjust TSS fixtures to be usable in tECDSA signing tests
The TSS fixtures we use for testing were copied from the TSS lib repository
when the TSS party key is built differently than in our code. We build the
TSS party key by just taking the integer value of the signing group member index
. In order to use those fixtures in our tECDSA signing tests, we need to adjust
the party keys stored in the `Ks` array and the `ShareID` that is actually
the key of the party owning the share.