Remove jobs veryfying contracts on Etherscan
We're switching to a different method of contracts verification on Etherscan.
Instad of veryfying contracts using `hardhat-deploy` plugin (which had problems
with verifying some contracts), we're doing that using
`@nomiclabs/hardhat-ethers` plugin. The use of new plugin was already added in a
different PR/commits. With that new plugin the contracts get verified during the
deployment step, this means that we don't need the
`contracts-etherscan-verification` job anymore in the CI workflows.
We're not removing the `hardhat-deploy` plugin, because it is still used for
contracts deployment.