Consistent party key between TSS runs
So far, we were doing a simple mapping between the `group.MemberIndex` and the
TSS party key for both tECDSA DKG and signing. Unfortunately, this turned to be
wrong. The TSS party key generated based on the `group.MemberIndex` has an
impact on the private key share produced as DKG result which means the same
TSS party key must be used by the share's holder during signing. However, our
DKG implementation contains a path where the final `group.MemberIndex` of a
member can differ from the original `group.MemberIndex` using during
protocol execution. That typically takes place when the DKG process completes
with misbehaved members that must be excluded from the final signing group.
In result, the `group.MemberIndex` <-> TSS party key mapping logic does not
work for the aforementioned case and makes signature generation impossible
if the signing group was generated with misbehaved members.

To make it work, we need to make sure that the key used to identify the party
during a TSS execution is preserved between runs. We achieve that by introducing
different mapping logic for DKG and signing.

For DKG, the TSS party key is now constructed by taking the `group.MemberIndex`
and adding it to the DKG seed. The seed component allows avoiding confusion
between the `group.MemberIndex` and party key and also ties the TSS keys to
a given DKG instance. Here is how it works for an example DKG run:

- Seed: 200
- Member indexes: [1, 2, 3, 4, 5]
- TSS keys: [201, 202, 203, 204, 205]

If member 3 misbehaves and gets excluded from the final signing group, the
member indexes and TSS keys arrays will look as follows:

- Member indexes: [1, 2, 4, 5]
- TSS keys: [201, 202, 204, 205]

The final member indexes must be shifted accordingly but TSS keys must remain
the same since they are connected with specific private key shares held by
members:

- Member indexes: [1, 2, 3, 4]
- TSS keys: [201, 202, 204, 205]

For signing, the member 1 must use key `201`, member 2 must use key `202`
and so on. Fortunately, each key share contains the full TSS keys list used
during DKG as reference. That information is held in the `Ks` field, and now
we are using it to map the `group.MemberIndex` to a proper TSS party key
during signing.

`