Extract proto generation to Makefile
We don't need to maintain multiple `gen.go` files for protobuf
generation. We defined a makefile's target for protobuf generation.

To run it execute `make gen_proto`.

We added a `go:generate` entry to the `main.go` so if someone is used to
using `go generate .` command the makefile's target will be executed.

Makefile gives advantage of regenerating the code only is the
prerequisites have changed, as prerequisites we defined `*.proto` files,
but also `go.mod` and `go.sum`, in case a dependency version is updated.