Replace github.com/gogo/protobuf in authenticate connection
We want to replace github.com/gogo/protobuf which is deprecated.
We want to use google.golang.org/protobuf/proto instead.

Unfortunately the new package doesn't expose `NewDelimitedReader` and
`NewDelimitedWriter`, so we had to change the code.

We were lucky that `google.golang.org/protobuf/proto` has support for
delimited read and writer already implemented, but this has not been
released yet, so we need to pin to a commit from master branch. We do
that only for the `protodelim` feature. The rest of imports will remaing
on the released version of google.golang.org/protobuf/proto
(v.1.28.1)