Tweak signing phases block duration to match benchmark results
According to the benchmark results (Apple M1 Max 10 cores), a single member
needs the following time to execute the given signing phase:

- TSS round 1: 2.08 s
- TSS round 2: 1.14 s
- TSS round 3: 0.76 s
- TSS round 4: 0.0002 s
- TSS round 5: 0.0100 s
- TSS round 6: 0.0004 s
- TSS round 7: 0.0200 s
- TSS round 8: 0.0001 s
- TSS round 9: 0.0010 s
- TSS finalization: 0.0003 s

If we take the worst case, i.e. 51 members executing signing on a single
machine, we assume 1 block is 12 seconds on average, and we add 20% of a
safety margin, we can compute the recommended count of active blocks for each
phase:

- TSS round 1: (51 * 2.0800 / 12) * 120% ≈ 11 blocks
- TSS round 2: (51 * 1.1400 / 12) * 120% ≈ 6 blocks
- TSS round 3: (51 * 0.7600 / 12) * 120% ≈ 4 blocks
- TSS round 4: (51 * 0.0002 / 12) * 120% ≈ 1 block
- TSS round 5: (51 * 0.0100 / 12) * 120% ≈ 1 block
- TSS round 6: (51 * 0.0004 / 12) * 120% ≈ 1 block
- TSS round 7: (51 * 0.0200 / 12) * 120% ≈ 1 block
- TSS round 8: (51 * 0.0001 / 12) * 120% ≈ 1 block
- TSS round 9: (51 * 0.0010 / 12) * 120% ≈ 1 block
- TSS finalization: (51 * 0.0003 / 12) * 120% ≈ 1

However, based on experiences from v1 DKG, we know that 1 active block
is often not enough in terms of networking, so we should use a higher amount.
The usual choice is 5 active blocks as this values gives compromise between
resiliency and execution time. That said, the final number of active blocks
for each phase is as follows:

- TSS round 1: 11 blocks
- TSS round 2: 6 blocks
- TSS round 3: 5 blocks
- TSS round 4: 5 blocks
- TSS round 5: 5 blocks
- TSS round 6: 5 blocks
- TSS round 7: 5 blocks
- TSS round 8: 5 blocks
- TSS round 9: 5 blocks
- TSS finalization: 5 blocks