Improved documentation of IdentityConverter
Added an explanation of how member index relates to TSS party ID during the
protocol execution. I am still not sure if the converter is the right place for
it but I think it's the best place we have now and this knowledge is helpful to
understand the need for the converter in the first place.