Make the 0-member-index case explicit in TssPartyIDToMemberIndex
There are two possible cases when Member Index can not be obtained from
party ID.

For dkg, it happens when seed < party ID.
For signing, it happens when there is no party ID with the key equal to
the member index.

These two cases are now explicit and MemberIndex(0) is expected to be
returned from TssPartyIDToMemberIndex.

In practice, it should never happen but it is better to be explicit
about corner cases.