Introduce keystore and work storage directories
We want to store generated preparameters in the storage. We figured out
that it should be separated from the keystore, as there will be a lot of
data stored, but the data are of a different kind. We introduce a `work`
subdirectory to store the data that are persistent to disk and should be
kept on disk for the client restarts but nothing bad happens when the
data is lost. This is opposed to the `keystore` directory that has to be
baced up with the greatest attention.