Trim excluded members for signing to the smallest possible count
The random retry algorithm for signing guarantees that the resulting list
will be at least as big as the honest threshold. For performance reasons,
we want to always use the count of members equal to the honest threshold
as more members are not needed for signing. Here we add the logic that performs
an additional exclusion round for signing.