Implement pre params persistence
We want to store the generated pre params on disk to use them after a
client restarts. Here we implement the persistent storage for pre
params.

Once the preparams are generated they are stored on disk in the storage
dir `work/preparams/` in a file named according to a template:
`pp_<CREATION TIMESTAMP MS>_<HASH 7 characters>`.

When the praparams are used by DKG they are deleted from the storage.